Rename FolioItem to Indexer::Item (and extract Indexer::CallNumber)
